delete old route_display and overwrite with new one
